property_service: only send properties that init is watching to init
Whenever a property is set, init checks to see if it matches any
property triggers or the current wait_for_prop property.  This is more
expensive now that property_service runs in a separate thread and
needs to inform init of each property change with a separate message.
This was noted in our boot timing dashboards.

This change informs property_service of which properties init is
watching and has property_service only send property changed messages
for them.

Bug: 141009024
Test: CF boots

Change-Id: I3d26005ad6ccccf6bdefbd21adb440760d9e218e