Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Ironically, I left this overnight hoping to find the same answer.
> > 
> > I'm assuming it has to do with this line from the other CL:
> > client->listKeysOfUid(KEY_PREFIX, KEY_UID, &aliases)
> > 
> > but I'm not sure why we need a new UID for this.  Can we list keys through a different mechanism than by UIDs?  Is there a different UID that could be used here if not?
> 
> Shawn and Janis, is it safe/preferred to just use prefix here (and system uid like others)?  I'd just like to avoid "cross-pollution" since the use cases should be mutually exclusive.

Keystore organizes keys by UID namespaces. In principle, the keys could be added to the AID_SYSTEM namespace with a unique prefix. However, the SYSTEM namespace is already fairly large, and I don't know if there is actually a good naming convention to avoid colisions. Above that the fs-verity keys need to be accessed by init (root) which would mean that we would have to open the SYSTEM namespace up to AID_ROOT. Following the principle of least authority, it would be good to have a separate namespace for this purpose.

Patch-set: 1