Update patch set 2
Patch Set 2: Verified+1

(3 comments)

Hi Tom, 

Thanks for the comments. The term DSU refers to the mechanism and GSI refers to specific images released by Google. We are working on supporting a DSU zip package which can contain arbitrary images as long as they are signed by the same OEM. The OTA package can be a special use case of this package where all update-able images are included.

Hi Bowgo, 
Thanks for the clarifications and the comments in BuildGsiFstabEntries. I fixed accordingly in #2. Currently we don't have a plan to make DSU partitions permanent under the DSU context. If the user like to make it permanent, one possible approach would be sending an intent to the update engine and have the virtual-ab to take over. In this scenario, the already created DSU partitions can be used like a cache. This part is beyond the scope though.

Patch-set: 2
Reviewer: Gerrit User 1135739 <1135739@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1