Create change
Uploaded patch set 1.

Patch-set: 1
Change-id: I492f7989ea8042912e9d0e0eadeaa351affbee13
Subject: This change is part of a topic that moves the recovery resources from the system partition to the vendor partition, if it exists, or the vendor directory on the system partition otherwise. The recovery resources are moving from the system image to the vendor partition so that a single system image may be used with either an A/B or a non-A/B vendor image. The topic removes a delta in the system image that prevented such reuse in the past.
Branch: refs/heads/master
Status: new
Topic: 
Commit: 26bc706295066272f0c524bd81da248448066c2c
Tag: autogenerated:gerrit:newPatchSet
Groups: 26bc706295066272f0c524bd81da248448066c2c
Private: false
Work-in-progress: false