Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (2 comments)
> 
> KISS duplication: use dependent patch w/o this one, adb remount is enough to trigger this every time.

I think the KISS plan here would be to have a new program:

int main() {
  int fd = posix_openpt(O_RDWR | O_CLOEXEC);
  close(fd);
  return 0;
}

And see if that dies due to SIGHUP or returns 0.  Depending on the result of that test, you could grow the program to include the rest of the function calls from here or debug directly a much smaller repro case.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>