logwrapper: move child_ptty open to child process
Opening child_ptty before fork() means that the parent process will
use child_ptty as the controlling terminal if it doesn't already have
one.  This is a problem, because when the parent_ptty closes, SIGHUP
will be sent to the parent process, since its controlling terminal has
closed.

It's better to have the child process start its own session and then
open child_ptty as its controlling terminal.  In this case, the child
process will get SIGHUP if the parent process exits, but the parent
process avoids the original issue.

Also add O_CLOEXEC where appropriate and restore the sigmask in the
parent, otherwise SIGINT and SIGQUIT will be blocked regardless of if
ignore_int_quit is specified.

Test: can launch processes through adb shell that use
      android_fork_execvp_ext() without problems

Change-Id: Ia76cd17e16535500170d8f5e2183b3bbbf843df7