Update patch set 1
Patch Set 1:

What about enabling it only for userdebug and eng builds then? If we're getting problems with rollover after that, the proper approach is arguably to reduce the default verbosity.

If not that, what can we do specifically avoid linker errors going to /dev/null for early boot binaries? If the linker is changed to send errors directly to kmsg, we'd still need selinux changes like the one in http://aosp/1127597, won't we?

Patch-set: 1
Reviewer: Gerrit User 1373864 <1373864@85c56323-6fa9-3386-8a01-6480fb634889>