Update patch set 3
Patch Set 3: Presubmit-Ready+1

(3 comments)

I agree with all that nnk@ says above.  I think it's fine to make this only an opt-in, especially with the goal of limiting its usage.

With regards to the linker, I don't have a great option here.  Perhaps we can append more information to the log when the exec() call fails to state if the error is "No such file or directory", then it could be a linker error and the linker should be debugged directly?

Truthfully, before your bug I've only seen one other case of a linker failure so early in boot.  It was when I was splitting 2nd stage init and adding SwitchRoot(), which few other people are ever going to be dealing with that level.  I think the additional error text may be sufficient.

Patch-set: 3
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1