Update patch set 6
Patch Set 6:

> Patch Set 6: Code-Review-1
> 
> why do we need a separate directory and Android.bp for vts? that still seems like a bug.

Agree that it does seem to be something we can improve. filed b/141684102
If the bug will take a while to resolve (I think the code change is simple, but just in case it takes too long in design), can we land this test first? We'll do the cleanup later.

Patch-set: 6
Reviewer: Gerrit User 1088076 <1088076@85c56323-6fa9-3386-8a01-6480fb634889>