Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1: Code-Review-1
> > > > 
> > > > No, AFAIK, this library is part of NDK API. It should be listed here.
> > > 
> > > I thought that apps are still able to load libneuralnetworks.so via "neuralnetworks" namespace. Then why should it be listed in public.libraries.txt?
> > 
> > We are still using the text file for other purposes: e.g. preloading.
> 
> I got it. I missed that libs in APEXes(art, neuralnetworks) are removed when loading public.libraries.txt. However, given that those libs are removed regardless preloading or not, I'm more curious why we still need to list them in public.libraries.txt. .. or did I miss something?

The preloading mechanism relies on public.libraries.txt. If you remove the lib from the text, the lib won't be preloaded. if might be okay to not preload libneuralnetworks.so because it is not a frequently used library, but that anyway is a regression.

Ideally, libnativeloader shouldn't remove a lib in an APEX from public.libraries.txt. Instead, it should detect that a lib in public.libraries.txt is from an APEX and then be able to redirect the link for the lib from the default namespace to the namespace for the APEX. Currently, libnativeloader can't do that because it lacks any knowledge about APEX. Everything about neuralnetwork APEX are hardcoded.

Patch-set: 1
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>