Update patch set 3
Patch Set 3:

Sorry for the late follow-up.

> The linker does not use liblog, it directly writes to stderr and the logd socket. 

I see, didn't realise that. It still links liblog statically though, so is there a particular reason it can't go through it too?

> I don't think that the problem is that we need a fallback to kmsg when logd isn't available, since logd is available extremely early, so that isn't a good enough signal.

logd is only started on init, isn't it? The case I had problems with was boringssl_self_test, which is on early-init. Looking at init.rc it seems there's quite a lot of stuff that happens before logd starts. Even though most of it doesn't involve exec'ing binaries at the moment, my worry is linker problems for those things that do, or will do in the future.

Or a bit more generally, getting coverage of any alog that take place in the early boot before logd is ready. I figure that like the linker, there's potential for running code in that stage that wasn't specifically written with the init use case in mind and hence isn't using kmsg directly.

> I think the real problem here is "if the system doesn't boot, can we still get the relevant logs from logd".

I agree that'd be nice too, but imo it's clearly secondary to the problem above. I'd like to be sure that I at least will get the logs locally with a serial cable.

> I filed b/141634082 to keep this in mind.

Noted, thanks!

> I still wonder, how many other people have been burnt by linker failures at such a very early stage. /.../

I've already argued this, but briefly: it did happen with the boringssl_self_test change, and given the type of changes we'll do with APEXes where various bits of the boot process is being moved out to separate binaries, I think it's likely to happen again. Hence past history might not be much to go at.

That said, I don't presume we got to fix it pronto based on the single bad experience I've had. I can just open a ticket to record the problem and we'll wait until the next time it hits.

> I believe this would be sufficient, especially since builds that are failing very early (before adb launches), already require a decent amount of hands on debugging.

The thing is that even knowing it's a failing exec and suspecting the linker, it is hard to get the alog messages at all. I only got the logs in the boringssl_self_test case by disabling reboot_on_failure and running it manually in the shell much later. I don't expect that to always be a possibility for similar boot problems. I suppose my next step would have been to redefine the logging macros in the linker to send the logs to kmsg directly, but doing that ad-hoc (and dealing with the inevitable selinux blockage on the way) adds to the debug time quite a bit.

Patch-set: 3
Reviewer: Gerrit User 1373864 <1373864@85c56323-6fa9-3386-8a01-6480fb634889>