Update patch set 3
Patch Set 3:

> Like I said, we can move starting logd earlier to cover these cases. /.../

If we can ensure it starts before code from APEXes gets running then it'd cover most of the use cases I've been arguing, yes.

It'd still leave a logging gap for the linker on earlier binaries (that don't use stdio_to_kmsg): init is fine, because everything logged in that process goes to kmsg, but e.g. linker debug log from the linking of logd will still be lost.

stdio_to_kmsg avoids that, so what about adding it to all the services that start before logd, and logd itself? The reason they shouldn't have it was that they're well-behaved and use kernel logging, wasn't it? But it would cover the linker log, and it presumably wouldn't hurt otherwise.

> Your issue was with the apex version of boringssl_self_test, right?  That starts much later than early-init. 

Indeed, was wrong on that.

> Those logs likely made it to logd, however since we reboot when it fails, there was no way to retrieve them.  This is why I think b/141634082 is the right solution here.

I hardcoded the reboots out of init itself so it wouldn't reboot no matter what, and I still didn't see any linker debug logs, so I concluded it was still before logd. But perhaps I did some other stupid mistake. Is it worthwhile for me to try to make a reproducible case with it?

Patch-set: 3
Reviewer: Gerrit User 1373864 <1373864@85c56323-6fa9-3386-8a01-6480fb634889>