Update patch set 1
Patch Set 1:

I don't think this does what you think it does.

The way QueueBuiltinAction() works is that it allocates a new Action instance and then puts an entry in the queue with the address of that Action.  Once that Action has been processed it's essentially impossible to queue it again, hence why we delete it.  A subsequent call can be made to QueueBuiltinAction() for the same function and it'll work normally; we do this for MixHwrngIntoLinuxRngAction.

There was a secondary intention too of making the actual actions from init.rc scripts oneshot to save on memory, but that never a materialized.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>