COW size calculator
Instead of using heuristics on the calculation of the COW size, use the
internal structure of the COW device as generated by the Linux kernel.
This simplifies the forecast of the required COW device size depending
on the write operations performed on the associated snapshot device.

Created PartitionCowCreatorTest::CowFileSize test.

Change-Id: If147001544988716430b36a4c75dfa7f34b4c8d9
Bug: 140835698
Test: libsnapshot_test
Signed-off-by: Alessio Balsini <balsini@google.com>