libsnapshot: init maps snapshot with less stacking
If a snapshot is mapped with force_writable = true, certain
ranges can be directly written (instead of being written to
dm-snapshot, which requires a bigger COW size). Hence, a
dm-linear device is created on top of the base and snapshot
device to distinguish regions that are directly written
or regions that are written through the snapshot.

During init, snapshots are mapped with force_writable = false,
so the dm-snapshot device can be returned directly.

Delete code that handles the stacking for merge and merge
complete (because we no longer have that stacking after reboot
any more).

Test: libsnapshot_test
Bug: 141889746
Change-Id: Ie92ba0ade8ef127159467829984cc7ff643382f7