Update patch set 7
Patch Set 7: Verified+1

> Patch Set 6:
> 
> it seems odd that init both spawns lmkd, and then init connects to lkmd using a different socket. Why doesn't init just create the socket, and pass the socket using standard file descriptor passing on exec? That avoids the attempts at retrying the connection and deals better with life cycle issues.

Creating a socket for lmkd using socketpair() would require some indication that lmkd is a special service that needs this kind of special treatment. The socket to communicate to lmkd already exists and AMS is already using it, so init can use it the same way even though lmkd is its child. After discussing this with Tom Cherry and considering additional complexity of introducing special case for lmkd he decided that we should not change the way we communicate with lmkd.

That said, this new patchset differs from the previous one quite drastically because I discovered some flaws in the previous design and added ability to:
- Check credentials of the lmkd client
- Prevent one client from unregistering or purging records of another client
- Unregister a process when it dies
- Re-register all services if lmkd died and got restarted

Patch-set: 7
Reviewer: Gerrit User 1157738 <1157738@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1