Add a test asserting that default fstab has /data entry
When aosp/1149619 is merged, userspace reboot will start to depend on
the fact that default fstab has an entry for /data, hence this CTS test
to make sure that this is the case.

I've added it to init_test for the lack of any better place. LMK if you
prefer to have a separate test file for this (e.g. fstab_test.cpp?)

Bug: 135984674
Test: atest CtsInitTestCases
Change-Id: I7a3c181a6b27a7b4dfc344ff04be712f3fe3376e