Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> > Patch Set 1:
> > 
> > (2 comments)
> > 
> > I think this can be even more helpful if it is implemented more centrally in libmodprobe. With the check in Insmod() we'll still end up iterating over a module's soft and hard dependencies in the library, before terminating in Insmod (assuming the module is loaded). If the check is put at the top of LoadWithAliases we can prevent that also.
> 
> Since one can call Insmod directly as well from the library, strictly we need the checks to be done in both places.

Do we have use cases where insmod is called directly from a process that loads this library?  The modprobe front end shouldn't be doing this, neither should first stage init, which are the two main (and only current) consumers.

If this is just for efficiency gains and none of our current users are otherwise calling insmod, then maybe best to defer this until we need it?  In any case, we'd probably want to provide them with a new insmod command instead of trying to hook the syscall.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>