Fix potential bad info in eh_frame_hdr.
Sometimes, an elf can have FDEs with pc_start and pc_end, while still
having another FDE for the same pc_start. The secondary FDE is the
one that should be used, and the empty FDE should be ignored. The problem
is that sometimes the eh_frame_hdr winds up pointing to the empty
FDE. In order to fix this, if the eh_frame_hdr points at an empty FDE
then try and find the real FDE directly from eh_frame.

The change cleans up and removes unused variables from DwarfEhFrameWithHdr
and changes the objects so that all of the DwarfSection objects and
DwarfEhFrameWithHdr object inherit from the same class.

Add new unit tests to verify this functionality.

Bug: 142483624

Test: Unit tests all pass.
Change-Id: I128a916e3ba378931de7d44ee15e57e24d4073df