Update patch set 1
Patch Set 1:

(11 comments)

A lot of comments but it all sounds good.  My biggest question is whether or not vendor init scripts were creating any of these top level directories and relied on the old behavior.  We'd need to keep it around for older vendor images if so.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>