simplify and better document checksum functions
Also introduce a little bit more consistency wrt. handling of 0 vs 0xFFFF,
which are kind of the same in 1's complement arithmentic:
0 is sometimes called positive 0, while 0xFFFF is sometimes called negative 0.

(one's complement 16-bit arithmentic is done modulo 0xFFFF, ie. 65535)

Per RFC, IPv4 and TCP checksums require 0 to be sent as 0, but accept 0xFFFF,
while UDPv4/UDPv6 checksums require 0 to be sent as 0xFFFF
(and special value of 0 means 'no checksum' for UDPv4 and is invalid for UDPv6,
unless used for tunneling, when it also means 'no checksum')

In practice Linux sometimes doesn't bother with the distinction
and just blindly uses 0xFFFF.  This is obviously correct for UDP,
and not quite according to spec for IPv4 and TCP, but works.

In paricular 'CHECKSUM_PARTIAL' checksum offload provides no way to
differentiate between UDP and TCP, and thus any nic utilizing tx csum offload,
must either do packet parsing or simply store a calculated checksum of 0
as 0xFFFF unconditionally.

Either way 'be lenient in what you accept' means 0xFFFF works for TCP receive.

Test: builds, atest
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I6886270a2f3282a3a977e5fb0f09253d85e8d4c5