Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> it's true of the whole library, not just this header... it's supposed
> to make the common cases hard to get wrong.
> 
> i thought we had a README.md with this and the "at least three
> unrelated projects that could be cleaned up" rule, but i see we don't.
> i'll send out a CL.

Also, if you want callers to bypass this property API but be compatible with the property-interpretation semantics here, we should expose the boolean-parsing "yes", "1", "no", etc., in the boolean-parsing code as a separate libbase parseBoolean function. Would you take a patch to do that? The alternative is to replicate this parsing logic, and replicated code can go out of sync.

Patch-set: 1