llkd: blacklist com.hbo.hbonow&[pm]
com.hba.hbonow forks a pm thread, which then dies and is unwaited for.
When a child process dies without being cleaned up, this is a signal
that the parent process is is not paying attention to standard cleanup
practices and quite possibly could accumulate and leak a lot of zombie
process, the parent process is locked up, or the kernel has a live
lock in the zombie process cleanup operations. Generally all are a
bad experience for the users.

Signed-off-by: Mark Salyzyn <salyzyn@google.com>
Bug: 144058139
Test: none
Change-Id: I112a7c505f2f23110ae972f3ca6259040eb39f7e