Restart bpfloader on zygote restart
bpfloader needs to restore the state of the system during Android
restart if the state ever got messed up. Such a mess up can cause the
system to never boot. To make it robust, lets restart the loader and
also reset the property so netd waits for a green signal from loader.

Bug: 143236645

Change-Id: I41129d5d7eba3fcde6b350c4e82a17275ae548b1
Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>