Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This is not a change that'll be accepted upstream. And this forces everyone to maintain a downstream MOD_ALIAS(). Do we really need this? Why not just load all modules in the ramdisk first before running any other code?

Upstream? There is nothing to upstream!

$ modprobe onEarlyInit || true
$ modprobe --all

This is specific to _init_ in Android.

As stated, certain modules do not load first naturally when built because of alphabetic/directory order. For example (one of several that are starting to show up), we _need_ the watchdog_v2 module to load _first_ before all other modules regardless of the order the the build has placed them into. Doing so gives us the best chance for capturing and diagnosing later modules that load that end up spinning until the hardware watchdog kicks in.

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0