Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > > This is not a change that'll be accepted upstream. And this forces everyone to maintain a downstream MOD_ALIAS(). Do we really need this? Why not just load all modules in the ramdisk first before running any other code?
> > 
> > Upstream? There is nothing to upstream!
> > 
> > $ modprobe onEarlyInit || true
> > $ modprobe --all
> > 
> > This is specific to _init_ in Android.
> > 
> > As stated, certain modules do not load first naturally when built because of alphabetic/directory order. For example (one of several that are starting to show up), we _need_ the watchdog_v2 module to load _first_ before all other modules regardless of the order the the build has placed them into. Doing so gives us the best chance for capturing and diagnosing later modules that load that end up spinning until the hardware watchdog kicks in.
> 
> Saravana was referring to the driver source,  which hopefully would get upstreamed. An Android-specific MODULE_ALIAS macro presumably would be frowned upon there and likely rejected...
> 
> But that aside, thinking about this a bit more, I'm not sure we need this after all. Vendors list the modules they want copied into the vendor ramdisk in BOARD_VENDOR_RAMDISK_KERNEL_MODULES. Although vendors may want to be able to use a wildcard to specify what goes into BOARD_VENDOR_KERNEL_MODULES and avoid listing them all out, the number of modules going into the ramdisk should be much lower. And the contents of BOARD_VENDOR_RAMDISK_KERNEL_MODULES is what goes into modules.load, dictating the load order. So they can just establish any desired ordering there.
> 
> That won't help someone trying to build with the kernel/build repo, so maybe we need to create a similar mechanism there.

I understand, I overstand and was leaping to the already established msm drivers as not one of them is at risk of going upstream (!). But the only repeating meme case is pure_initcall (not available to modules) which also happen to be the modules that need to be early-loaded at priority thus far. I have not seen any abuse of this call except for the modules that also happen to need to be loaded before any of the later modules. Thus the code change is isolated to vendor code modifications of upstream. upstream any sources with pure_initcall by definition can not be modularized without some efforts, and to allow it to be modularized carries compromises too that make it difficult to be taken upstream (the worst case is the early-init serial console driver).

BOARD_VENDOR_RAMDISK_KERNEL_MODULES will work if order is never $(sort), but as you state, does not work for a kernel build repo :-(

By all means create a similar mechanism, perhaps the kernel repo _needs_ to contain the /vendor or /device trees to pick this variable up; or the kernel repo (or rather the products landed in /device/<manufacturer>/<product>_kernel) should be the one _feeding_ the values in BOARD_VENDOR_RAMDISK_KERNEL_MODULES? One source of truth needs to be established, so we are never out of sync.

My proposal is a coarse means for the as-built driver modules to specify that they need to be loaded at priority, and so far I have found few reasons (camera driver was one) where the order needed to be fine tuned.

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0