Define NO_LIBDEXFILE_SUPPORT for all vendor and recovery variants
Previously, NO_LIBDEXFILE_SUPPORT was defined only when libbacktrace and
libunwindstack are built. As a result, a vendor library using the
two libraries is built with headers that are different from the headers
that were used to build the two libraries.

Fixing this problem by defining NO_LIBDEXFILE_SUPPORT when the context
is __ANDROID_VNDK__ or __ANDROID_RECOVERY__.

Bug: 144470551
Test: m
Change-Id: I2307f301721fb2213e6abc6fbe38fbd0b0ff2ef9