logd: don't skip log entries for slow readers, kick them instead
If log readers cannot read entries as fast as they are being
generated, then we will run into a situation where log buffers are
consuming more memory than intended.

We do not want to block log writers, since that would cause system
wide performance issues.

We do not want to let memory growing indefinitely, so we limit the max
growth to 2x the intended buffer size, then we kick out any client
that is causing these log messages to be retained.

Previously, we would first skip over log messages in the log buffer,
but that is ambiguous to readers.  We prefer to send a clear
indication that the logger is too slow and will not receive all logs,
by disconnecting it.

Bug: 144311420
Test: TBD
Change-Id: Idece01b395b9d928cd2b45e04f349106d7d04b0a