libsnapshot: split lock file and state file
We used to flock() on /metadata/ota/state to ensure
atomic access. However, writing the file itself is
not necessary atomic and may lead to inconsistent
states.

This change creates a separate file, /metadata/ota/lock.
flock() is called on this file instead of /state. This allows
a follow-up change to turn all writes to the /metadata partition
atomic.

Test: libsnapshot_test
Test: apply OTA from older build to this, then reboot

Bug: 144549076

Change-Id: Ib4dd9e9be1a43013c328e181b9398ac0b514dbce