Update patch set 7
Patch Set 7:

I'm rather hesitant here.  Do we have any precedent of mounting over a directory in /system?  I don't think so and I don't think it would be a good idea to start.

I also don't fully understand what the problem is with the runtime tests; both this CL and the bug are sparse on the description of the actual issue here.  How do the tests call chroot() and how does the chroot cause this issue?  Can either here or the bug have a link to the test in question or more details about what is going wrong?  I'd like to see if there are other ways to solve this issue.

Patch-set: 7
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>